Add Enum descriptions to Show All Values dropdown
This commit fixes several issues present such as accessing nonexisting
keys when attempting to display descriptions and using conditional logic
in the template, which causes inconsistent UI. There was also no built-in
logic to handle displaying enum variables, which was expected. Part of
refactoring this code included some changes to the template logic to
use consistent UI between enums with descriptions and enums without
descriptions. Another UI feature added was a fade near the show all
values button to give visual indicator to a user that there was more
information in the accordion if they clicked the button.

Some further work may be to change the styling of the enum value and
its associated variables to be more emphasized, as they currently have
much less contrast against the background than is ideal, at least in dark
mode.

 - There is no longer any separation in UI between enums with
descriptions and enums without descriptions.

 - This commit builds in CSS and functions for processing enum
variables similarly to the way enum descriptions are processed.

Resolves: #4