Fixed Additional Properties Render
The "additional properties" button used to appear alongside every
property rendering all at once. The issue was to do with the
DisclosurePanel's static field evaluated to true, thus disregarding the
hiding functionality for the additional properties. This is the more
straightforward approach to solving this issue without doing large
refactors on template logic; the DisclosurePanel logic was flawed.

Resolves #12