Replace btoa/atob with js-base64 encoding
This commit replaces the JavaScript btoa() Base64 encoding function with
 js-base64's encode() function. The js-base64 function supports encoding
 non-Latin1 Unicode characters into Base64, thereby increasing
accessibility for user credentials. This change affects both
BasicAuthentication and OAuth2, which both used the built-in btoa()
and atob() functions that do not support non-Latin1 Unicode characters.
Additionally, a test case for non-Latin1 Unicode characters in user
credentials was added to ensure that the accessibility changes are
maintained going forward.

Resolves: #13