Enforce color theme and add initial config logic
Logic was added to enforce a consistent color theming layout based upon
useColorMode fallbacks. If darkMode is not provided, then look at
localStorage for the theme, and if that isn't present either, look at browser
preferences. The reason this was added is to enforce consistency and
clearly define what theme will be used each time a developer loads the
API reference. Some updates were made to the buttons as well to make
certain that their visual state matches the visual theme of the page, i.e.
if it is dark, then the toggle shows the moon, and light is the sun icon.

Resolves: #17